requirements.txt
Flask==3.1.0
gunicorn==23.*
python-dotenv==1.*
SQLAlchemy==2.0.39
PyMySQL==1.1.1
alembic==1.15.1